Move all post-related components into one folder
