+++ src/components/post/post-attachment-audio.jsx
@@ -2,8 +2,8 @@ import { PureComponent } from 'react';
 import { faFileAudio } from '@fortawesome/free-regular-svg-icons';
 import { faTimes } from '@fortawesome/free-solid-svg-icons';
 
-import { formatFileSize } from '../utils';
-import { Icon } from './fontawesome-icons';
+import { formatFileSize } from '../../utils';
+import { Icon } from '../fontawesome-icons';
 
 class AudioAttachment extends PureComponent {
   handleClickOnRemoveAttachment = () => {
