+++ src/components/post/post-attachment-general.jsx
@@ -2,8 +2,8 @@ import { PureComponent } from 'react';
 import { faFile } from '@fortawesome/free-regular-svg-icons';
 import { faTimes } from '@fortawesome/free-solid-svg-icons';
 
-import { formatFileSize } from '../utils';
-import { Icon } from './fontawesome-icons';
+import { formatFileSize } from '../../utils';
+import { Icon } from '../fontawesome-icons';
 
 class GeneralAttachment extends PureComponent {
   handleClickOnRemoveAttachment = () => {
