+++ src/components/post/post-attachment-image.jsx
@@ -2,8 +2,8 @@ import { PureComponent } from 'react';
 import classnames from 'classnames';
 import { faTimes } from '@fortawesome/free-solid-svg-icons';
 
-import { formatFileSize } from '../utils';
-import { Icon } from './fontawesome-icons';
+import { formatFileSize } from '../../utils';
+import { Icon } from '../fontawesome-icons';
 
 class PostAttachmentImage extends PureComponent {
   handleRemoveImage = () => {
