+++ src/components/post/post-comment-likes.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/helvetica/functions.scss';
+@import '../../../styles/helvetica/functions.scss';
 
 .container {
   display: flex;
@@ -36,9 +36,11 @@
   display: flex;
   align-items: flex-start;
   width: 240px;
+
   & > *:first-child {
     margin-right: 10px;
   }
+
   margin-right: 15px;
   margin-bottom: 15px;
 }
