+++ src/components/post/post-comments/collapse-comments.jsx
@@ -2,8 +2,8 @@ import { useEffect, useRef, useState } from 'react';
 import cn from 'classnames';
 import { faChevronUp } from '@fortawesome/free-solid-svg-icons';
 
-import { Icon } from '../fontawesome-icons';
-import { ButtonLink } from '../button-link';
+import { Icon } from '../../fontawesome-icons';
+import { ButtonLink } from '../../button-link';
 
 const observersSupported =
   typeof IntersectionObserver !== 'undefined' && typeof MutationObserver !== 'undefined';
