+++ test/jest/post-more-menu.test.jsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 import { createStore } from 'redux';
 import * as reactRedux from 'react-redux';
 
-import { PostMoreMenu } from '../../src/components/post-more-menu';
+import { PostMoreMenu } from '../../src/components/post/post-more-menu';
 
 jest.mock('../../src/components/fontawesome-icons', () => ({
   Icon: ({ icon }) => `fontawesome icon ${icon.iconName}`, // mocking out icon to make snapshots smaller
